requirements.txt
google-generativeai>=0.8.0
langchain>=0.2.0
langchain-google-genai>=0.1.0
pydantic[email]>=2.7.0
phonenumbers>=8.13.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
streamlit>=1.37.0